feat: #489 React hook refactor
