+++ apps/web/eslint.config.mjs
@@ -1,10 +1,32 @@
 import { defineConfig, globalIgnores } from "eslint/config";
+import importPlugin from "eslint-plugin-import";
 import nextVitals from "eslint-config-next/core-web-vitals";
 import nextTs from "eslint-config-next/typescript";
 
 const eslintConfig = defineConfig([
   ...nextVitals,
   ...nextTs,
+  {
+    plugins: {
+      import: importPlugin,
+    },
+    rules: {
+      "import/no-restricted-paths": [
+        "error",
+        {
+          basePath: ".",
+          zones: [
+            {
+              target: "./**/*",
+              from: ["./utils/use[A-Z]*", "./utils/authLoginFormSchema.*"],
+              message:
+                "Move React hooks and validation schemas out of utils/ into hooks/ or lib/validation/.",
+            },
+          ],
+        },
+      ],
+    },
+  },
   // Override default ignores of eslint-config-next.
   globalIgnores([
     // Default ignores of eslint-config-next:
