+++ apps/web/hooks/form/useLoginFormSchema.ts
@@ -2,7 +2,7 @@
 
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
-import { createLoginFormSchema } from "./authLoginFormSchema";
+import { createLoginFormSchema } from "@/lib/validation/auth";
 
 export function useLoginFormSchema() {
   const { t } = useTranslation();
