+++ apps/web/hooks/landing/useSlideData.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useTranslation } from "react-i18next";
 import { slideImages } from "@/utils/data";
